atualização folly field

+++ lib/crud/abstract_list.dart
@@ -43,13 +43,20 @@ abstract class AbstractList<
   final int itemsPerPage;
   final int qtdSuggestions;
   final List<AbstractRoute> actionRoutes;
-  final Future<Widget> Function(
+  final Future<Widget?> Function(
     BuildContext context,
     T model,
     UI uiBuilder,
     C consumer,
     bool edit,
   )? onLongPress;
+  final Map<AbstractRoute,
+      Future<bool> Function(BuildContext context, T model)>? actionFunctions;
+  final String? searchFieldLabel;
+  final TextStyle? searchFieldStyle;
+  final InputDecorationTheme? searchFieldDecorationTheme;
+  final TextInputType? searchKeyboardType;
+  final TextInputAction searchTextInputAction;
 
   ///
   ///
@@ -76,7 +83,14 @@ abstract class AbstractList<
     this.qtdSuggestions = 15,
     this.actionRoutes = const <AbstractRoute>[],
     this.onLongPress,
-  }) : super(key: key);
+    this.actionFunctions,
+    this.searchFieldLabel,
+    this.searchFieldStyle,
+    this.searchFieldDecorationTheme,
+    this.searchKeyboardType,
+    this.searchTextInputAction = TextInputAction.search,
+  })  : assert(searchFieldStyle == null || searchFieldDecorationTheme == null),
+        super(key: key);
 
   ///
   ///
@@ -121,6 +135,9 @@ class _AbstractListState<
 
   final Map<String, String> _qsParam = <String, String>{};
 
+  final Map<ConsumerPermission, AbstractRoute> permissions =
+      <ConsumerPermission, AbstractRoute>{};
+
   ///
   ///
   ///
@@ -141,7 +158,27 @@ class _AbstractListState<
       }
     });
 
-    _loadData(context);
+    _loadPermissions(context);
+  }
+
+  ///
+  ///
+  ///
+  Future<void> _loadPermissions(BuildContext context) async {
+    if (widget.actionFunctions != null) {
+      for (MapEntry<AbstractRoute,
+              Future<bool> Function(BuildContext context, T model)> entry
+          in widget.actionFunctions!.entries) {
+        AbstractRoute route = entry.key;
+
+        ConsumerPermission permission =
+            await widget.consumer.checkPermission(context, route.routeName);
+
+        if (permission.view) permissions[permission] = route;
+      }
+    }
+
+    await _loadData(context);
   }
 
   ///
@@ -301,6 +338,12 @@ class _AbstractListState<
                       itemsPerPage: widget.itemsPerPage,
                       uiBuilder: widget.uiBuilder,
                       consumer: widget.consumer,
+                      searchFieldLabel: widget.searchFieldLabel,
+                      searchFieldStyle: widget.searchFieldStyle,
+                      searchFieldDecorationTheme:
+                          widget.searchFieldDecorationTheme,
+                      keyboardType: widget.searchKeyboardType,
+                      textInputAction: widget.searchTextInputAction,
                     ),
                   ).then((T? entity) {
                     if (entity != null) {
@@ -352,10 +395,10 @@ class _AbstractListState<
                                 onPressed: () =>
                                     Navigator.of(context).pushNamed(route.path),
                               )
-                            : Container();
+                            : Container(width: 0, height: 0);
                       }
 
-                      return Container();
+                      return Container(width: 0, height: 0);
                     },
                   ),
                 );
@@ -509,8 +552,42 @@ class _AbstractListState<
       ),
       title: widget.uiBuilder.getTitle(model),
       subtitle: widget.uiBuilder.getSubtitle(model),
-      trailing: canDelete
-          ? IconButton(
+      trailing: Row(
+        mainAxisSize: MainAxisSize.min,
+        mainAxisAlignment: MainAxisAlignment.end,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: <Widget>[
+          ...permissions.entries.map(
+            (MapEntry<ConsumerPermission, AbstractRoute> entry) {
+              ConsumerPermission permission = entry.key;
+              // TODO - Create an Action Route component.
+              return FutureBuilder<bool>(
+                initialData: false,
+                future: widget.actionFunctions![entry.value]!(context, model),
+                builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
+                  if (snapshot.hasData && snapshot.data!) {
+                    return IconButton(
+                      tooltip: permission.name,
+                      icon: IconHelper.faIcon(permission.iconName),
+                      onPressed: () async {
+                        dynamic refresh = await Navigator.of(context).pushNamed(
+                          entry.value.path,
+                          arguments: model,
+                        );
+
+                        if (refresh is bool && refresh) {
+                          await _loadData(context, clear: true);
+                        }
+                      },
+                    );
+                  }
+                  return Container(width: 0, height: 0);
+                },
+              );
+            },
+          ),
+          if (canDelete)
+            IconButton(
               icon: Icon(FontAwesomeIcons.trashAlt),
               onPressed: () async {
                 bool refresh = await _deleteEntity(model, ask: true);
@@ -518,8 +595,9 @@ class _AbstractListState<
                   await afterDeleteRefresh();
                 }
               },
-            )
-          : Container(width: 1, height: 1),
+            ),
+        ],
+      ),
       onTap: onTap != null
           ? () => onTap(model)
           : () => _internalRoute(model, !selection),
@@ -682,7 +760,39 @@ class InternalSearch<
     required this.qsParam,
     required this.forceOffline,
     required this.itemsPerPage,
-  });
+    required String? searchFieldLabel,
+    required TextStyle? searchFieldStyle,
+    required InputDecorationTheme? searchFieldDecorationTheme,
+    required TextInputType? keyboardType,
+    required TextInputAction textInputAction,
+  }) : super(
+          searchFieldLabel: searchFieldLabel,
+          searchFieldStyle: searchFieldStyle,
+          searchFieldDecorationTheme: searchFieldDecorationTheme,
+          keyboardType: keyboardType,
+          textInputAction: textInputAction,
+        );
+
+  ///
+  ///
+  ///
+  @override
+  ThemeData appBarTheme(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    return theme.copyWith(
+      appBarTheme: AppBarTheme(
+        brightness: theme.colorScheme.brightness,
+        backgroundColor: theme.colorScheme.surface,
+        iconTheme: theme.primaryIconTheme.copyWith(color: Colors.grey),
+        textTheme: theme.textTheme,
+      ),
+      inputDecorationTheme: searchFieldDecorationTheme ??
+          InputDecorationTheme(
+            hintStyle: searchFieldStyle ?? theme.inputDecorationTheme.hintStyle,
+            border: InputBorder.none,
+          ),
+    );
+  }
 
   ///
   ///
@@ -853,7 +963,8 @@ class InternalSearch<
                                 'Sugestões:',
                                 style: TextStyle(
                                   fontStyle: FontStyle.italic,
-                                  color: Theme.of(context).accentColor,
+                                  color:
+                                      Theme.of(context).colorScheme.secondary,
                                 ),
                               ),
                             ),
